+++ contrib/localnet/orchestrator/start-zetae2e.sh
@@ -10,6 +10,8 @@
 # Trap signals and forward to children
 trap 'kill -- -$$' SIGINT SIGTERM
 
+/usr/sbin/sshd
+
 get_zetacored_version() {
   retries=10
   node_info=""
@@ -112,6 +114,9 @@ fund_eth_from_config '.additional_accounts.user_bitcoin_withdraw.evm_address' 10
 
 # unlock solana tester accounts
 fund_eth_from_config '.additional_accounts.user_solana.evm_address' 10000 "solana tester"
+
+# unlock sui tester accounts
+fund_eth_from_config '.additional_accounts.user_sui.evm_address' 10000 "sui tester"
 
 # unlock miscellaneous tests accounts
 fund_eth_from_config '.additional_accounts.user_misc.evm_address' 10000 "misc tester"