+++ pkg/chains/chain.go
@@ -56,13 +56,14 @@ func (chain Chain) IsExternalChain() bool {
 // on EVM chain, it is 20Bytes
 // on Bitcoin chain, it is P2WPKH address, []byte(bech32 encoded string)
 func (chain Chain) EncodeAddress(b []byte) (string, error) {
-	switch chain.Consensus {
-	case Consensus_ethereum:
+	if chain.Vm == Vm_evm {
 		addr := ethcommon.BytesToAddress(b)
 		if addr == (ethcommon.Address{}) {
 			return "", fmt.Errorf("invalid EVM address")
 		}
 		return addr.Hex(), nil
+	}
+	switch chain.Consensus {
 	case Consensus_bitcoin:
 		addrStr := string(b)
 		chainParams, err := GetBTCChainParams(chain.ChainId)
@@ -134,8 +135,9 @@ func DecodeAddressFromChainID(chainID int64, addr string, additionalChains []Cha
 		if err != nil {
 			return nil, fmt.Errorf("invalid TON address %q: %w", addr, err)
 		}
-
 		return []byte(acc.ToRaw()), nil
+	case IsSuiChain(chainID, additionalChains):
+		return []byte(addr), nil
 	default:
 		return nil, fmt.Errorf("chain (%d) not supported", chainID)
 	}