+++ zetaclient/chains/solana/signer/signer.go
@@ -132,6 +132,7 @@ func (signer *Signer) TryProcessOutbound(
 	coinType := cctx.InboundParams.CoinType
 
 	var tx *solana.Transaction
+	var fallbackTx *solana.Transaction
 
 	switch coinType {
 	case coin.CoinType_Cmd:
@@ -150,8 +151,14 @@ func (signer *Signer) TryProcessOutbound(
 				logger.Error().Err(err).Msgf("TryProcessOutbound: Fail to sign execute outbound")
 				return
 			}
+			incrementNonceTx, err := signer.prepareIncrementNonceTx(ctx, cctx, height, logger)
+			if err != nil {
+				logger.Error().Err(err).Msgf("TryProcessOutbound: Fail to sign increment_nonce outbound")
+				return
+			}
 
 			tx = executeTx
+			fallbackTx = incrementNonceTx
 		} else {
 			withdrawTx, err := signer.prepareWithdrawTx(ctx, cctx, height, logger)
 			if err != nil {
@@ -170,7 +177,14 @@ func (signer *Signer) TryProcessOutbound(
 				return
 			}
 
+			incrementNonceTx, err := signer.prepareIncrementNonceTx(ctx, cctx, height, logger)
+			if err != nil {
+				logger.Error().Err(err).Msgf("TryProcessOutbound: Fail to sign increment_nonce outbound")
+				return
+			}
+
 			tx = executeSPLTx
+			fallbackTx = incrementNonceTx
 		} else {
 			withdrawSPLTx, err := signer.prepareWithdrawSPLTx(ctx, cctx, height, logger)
 			if err != nil {
@@ -196,13 +210,51 @@ func (signer *Signer) TryProcessOutbound(
 	signer.SetRelayerBalanceMetrics(ctx)
 
 	// broadcast the signed tx to the Solana network
-	signer.broadcastOutbound(ctx, tx, chainID, nonce, logger, zetacoreClient)
+	signer.broadcastOutbound(ctx, tx, fallbackTx, chainID, nonce, logger, zetacoreClient)
+}
+
+// signTx creates and signs solana tx containing provided instruction with relayer key.
+func (signer *Signer) signTx(ctx context.Context, inst *solana.GenericInstruction) (*solana.Transaction, error) {
+	// get a recent blockhash
+	recent, err := signer.client.GetLatestBlockhash(ctx, rpc.CommitmentFinalized)
+	if err != nil {
+		return nil, errors.Wrap(err, "getLatestBlockhash error")
+	}
+
+	// create a transaction that wraps the instruction
+	tx, err := solana.NewTransaction(
+		[]solana.Instruction{
+			// TODO: outbound now uses 5K lamports as the fixed fee, we could explore priority fee and compute budget
+			// https://github.com/zeta-chain/node/issues/2599
+			// programs.ComputeBudgetSetComputeUnitLimit(computeUnitLimit),
+			// programs.ComputeBudgetSetComputeUnitPrice(computeUnitPrice),
+			inst},
+		recent.Value.Blockhash,
+		solana.TransactionPayer(signer.relayerKey.PublicKey()),
+	)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to create new tx")
+	}
+
+	// relayer signs the transaction
+	_, err = tx.Sign(func(key solana.PublicKey) *solana.PrivateKey {
+		if key.Equals(signer.relayerKey.PublicKey()) {
+			return signer.relayerKey
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, errors.Wrap(err, "signer unable to sign transaction")
+	}
+
+	return tx, nil
 }
 
 // broadcastOutbound sends the signed transaction to the Solana network
 func (signer *Signer) broadcastOutbound(
 	ctx context.Context,
 	tx *solana.Transaction,
+	fallbackTx *solana.Transaction,
 	chainID int64,
 	nonce uint64,
 	logger zerolog.Logger,
@@ -247,6 +299,11 @@ func (signer *Signer) broadcastOutbound(
 			rpc.TransactionOpts{PreflightCommitment: rpc.CommitmentProcessed},
 		)
 		if err != nil {
+			// in case it is not failure due to nonce mismatch, replace tx with fallback tx
+			// probably need a better way to do this, but currently this is the only error to tolerate like this
+			if !strings.Contains(err.Error(), "NonceMismatch") {
+				tx = fallbackTx
+			}
 			logger.Warn().Err(err).Fields(lf).Msgf("SendTransactionWithOpts failed")
 			backOff *= 2
 			continue
@@ -257,6 +314,43 @@ func (signer *Signer) broadcastOutbound(
 		signer.reportToOutboundTracker(ctx, zetacoreClient, chainID, nonce, txSig, logger)
 		break
 	}
+}
+
+func (signer *Signer) prepareIncrementNonceTx(
+	ctx context.Context,
+	cctx *types.CrossChainTx,
+	height uint64,
+	logger zerolog.Logger,
+) (*solana.Transaction, error) {
+	params := cctx.GetCurrentOutboundParam()
+	// compliance check
+	cancelTx := compliance.IsCctxRestricted(cctx)
+	if cancelTx {
+		compliance.PrintComplianceLog(
+			logger,
+			signer.Logger().Compliance,
+			true,
+			signer.Chain().ChainId,
+			cctx.Index,
+			cctx.InboundParams.Sender,
+			params.Receiver,
+			"SOL",
+		)
+	}
+
+	// sign gateway increment_nonce message by TSS
+	msg, err := signer.createAndSignMsgIncrementNonce(ctx, params, height, cancelTx)
+	if err != nil {
+		return nil, err
+	}
+
+	// sign the increment_nonce transaction by relayer key
+	inst, err := signer.createIncrementNonceInstruction(*msg)
+	if err != nil {
+		return nil, errors.Wrap(err, "error creating increment nonce instruction")
+	}
+
+	return signer.signTx(ctx, inst)
 }
 
 func (signer *Signer) prepareWithdrawTx(
@@ -288,12 +382,12 @@ func (signer *Signer) prepareWithdrawTx(
 	}
 
 	// sign the withdraw transaction by relayer key
-	tx, err := signer.signWithdrawTx(ctx, *msg)
+	inst, err := signer.createWithdrawInstruction(*msg)
 	if err != nil {
-		return nil, errors.Wrap(err, "signWithdrawTx error")
+		return nil, errors.Wrap(err, "error creating withdraw instruction")
 	}
 
-	return tx, nil
+	return signer.signTx(ctx, inst)
 }
 
 func (signer *Signer) prepareExecuteTx(
@@ -351,12 +445,12 @@ func (signer *Signer) prepareExecuteTx(
 	}
 
 	// sign the execute transaction by relayer key
-	tx, err := signer.signExecuteTx(ctx, *msgExecute)
+	inst, err := signer.createExecuteInstruction(*msgExecute)
 	if err != nil {
-		return nil, errors.Wrap(err, "signExecuteTx error")
+		return nil, errors.Wrap(err, "error creating execute instruction")
 	}
 
-	return tx, nil
+	return signer.signTx(ctx, inst)
 }
 
 func (signer *Signer) prepareWithdrawSPLTx(
@@ -401,12 +495,12 @@ func (signer *Signer) prepareWithdrawSPLTx(
 	}
 
 	// sign the withdraw transaction by relayer key
-	tx, err := signer.signWithdrawSPLTx(ctx, *msg)
+	inst, err := signer.createWithdrawSPLInstruction(*msg)
 	if err != nil {
-		return nil, errors.Wrap(err, "signWithdrawSPLTx error")
+		return nil, errors.Wrap(err, "error creating withdraw SPL instruction")
 	}
 
-	return tx, nil
+	return signer.signTx(ctx, inst)
 }
 
 func (signer *Signer) prepareExecuteSPLTx(
@@ -473,12 +567,12 @@ func (signer *Signer) prepareExecuteSPLTx(
 	}
 
 	// sign the execute spl transaction by relayer key
-	tx, err := signer.signExecuteSPLTx(ctx, *msgExecuteSpl)
+	inst, err := signer.createExecuteSPLInstruction(*msgExecuteSpl)
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "error creating execute SPL instruction")
 	}
 
-	return tx, nil
+	return signer.signTx(ctx, inst)
 }
 
 func (signer *Signer) prepareWhitelistTx(
@@ -510,12 +604,12 @@ func (signer *Signer) prepareWhitelistTx(
 	}
 
 	// sign the whitelist transaction by relayer key
-	tx, err := signer.signWhitelistTx(ctx, msg)
+	inst, err := signer.createWhitelistInstruction(msg)
 	if err != nil {
-		return nil, errors.Wrap(err, "signWhitelistTx error")
+		return nil, errors.Wrap(err, "error creating whitelist instruction")
 	}
 
-	return tx, nil
+	return signer.signTx(ctx, inst)
 }
 
 func (signer *Signer) decodeMintAccountDetails(ctx context.Context, asset string) (token.Mint, error) {