+++ zetaclient/testutils/mocks/sui_gen.go
@@ -18,6 +18,7 @@ type suiClient interface {
 	HealthCheck(ctx context.Context) (time.Time, error)
 	GetLatestCheckpoint(ctx context.Context) (models.CheckpointResponse, error)
 	QueryModuleEvents(ctx context.Context, q client.EventQuery) ([]models.SuiEventResponse, string, error)
+	GetOwnedObjectID(ctx context.Context, ownerAddress, structType string) (string, error)
 
 	SuiXGetReferenceGasPrice(ctx context.Context) (uint64, error)
 	SuiXQueryEvents(ctx context.Context, req models.SuiXQueryEventsRequest) (models.PaginatedEventsResponse, error)
@@ -25,5 +26,10 @@ type suiClient interface {
 	SuiGetTransactionBlock(
 		ctx context.Context,
 		req models.SuiGetTransactionBlockRequest,
+	) (models.SuiTransactionBlockResponse, error)
+	MoveCall(ctx context.Context, req models.MoveCallRequest) (models.TxnMetaData, error)
+	SuiExecuteTransactionBlock(
+		ctx context.Context,
+		req models.SuiExecuteTransactionBlockRequest,
 	) (models.SuiTransactionBlockResponse, error)
 }